+++ sw-config.js
@@ -35,7 +35,8 @@ const config = {
 	stale: [
 		/* Scripts */
 		'/js/index.min.js',
-		'https://cdn.kernvalley.us/js/shims.min.js',
+		'https://unpkg.com/@shgysk8zer0/polyfills@0.0.5/all.min.js',
+		'https://unpkg.com/@shgysk8zer0/kazoo@0.0.5/harden.js',
 		'https://cdn.kernvalley.us/js/security.min.js',
 		'https://cdn.kernvalley.us/components/leaflet/map.min.js',