+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Implements import maps
 - Add support and enforcement of `TrustedTypes`
 
 ## [v2.3.1] - 2023-01-21